requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.10
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
